refactor(refs T30969): move DpMultiselect to components root
