+++ src/components/DpMultiselect/DpMultiselect.stories.mdx
@@ -1,3 +1,10 @@
+import { Meta } from '@storybook/addon-docs'
+import DpMultiselect from './DpMultiselect'
+
+<Meta
+  title="Components/Multiselect"
+  component={DpMultiselect} />
+
 # Multiselect
 
 This is the generic multiselect functional component for vue instances.
@@ -18,6 +25,12 @@ Component props:
 Groups in select: The options list can also contain groups. It requires passing 3 additional props: group-label, group-values and group-select. group-label is used to locate the group label. group-values should point to the group’s option list. group-select is used to define if selecting the group label should select/unselect all values in the group, or do nothing. Despite that the available options are grouped, the selected options are stored as a flat array of objects.
 
 For all possible props see the [vue-multiselect docs](https://vue-multiselect.js.org/#sub-props).
+
+export const Template = (args, { argTypes }) => ({
+  props: Object.keys(argTypes),
+  components: { DpMultiselect },
+  template: `<dp-multiselect v-bind="$props" />`
+});
 
 ## Complete example
